% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preview.R
\name{mod_preview}
\alias{mod_preview}
\title{Launch the Wordcloud Shiny App}
\usage{
mod_preview()
}
\description{
This function launches the Shiny app for the wordcloud module.
It provides a user interface for generating word clouds from text data.
}
