% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requal_module.R
\name{mod_server}
\alias{mod_server}
\title{Wordcloud Server Function}
\usage{
mod_server(id, api)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}

\item{api}{An instance of the RequalAPI class.}
}
\description{
Wordcloud Server Function
}
