% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requal_module.R
\name{mod_server}
\alias{mod_server}
\title{wordcloud Server Functions}
\usage{
mod_server(id, api)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{con}{Connection object for Requal database.}

\item{user_id}{ID of the Requal user.}
}
\description{
wordcloud Server Functions
}
